The compiler is in a strange state at the moment. I have attempted to add initial parsing support for simple declarations, but it isn't yet working. At the moment we have somewhat lackluster support for declaring global integer types with some storage class specifiers thrown in.  The only concrete achievement I have for this commit is that I fixed a bug in the lexer where it would continue to return new tokens even after the end of file had been reached.
Furthermore, there appears to be a bug in the way that the parser uses the
lexer that causes the source code line count to get messed up. Testing the
lexer by itself yields the right results, but when I test the parser, the line
count presented at the end is wrong by 3 or more lines (too few lines).

The next step will probably be to write a symbol table and accompanying
routines in order to get the declarations we see installed properly. After
that we will move on to function declarations, definitions, and the simplest
of expression parsing.