Fixed a bug in gettok and peektok that would cause peektok to yield the wrong results when several gettok's had been used in a row without peeking. Peektok now returns the next token in the stream without advancing it, as it is supposed to. (In truth it does advance the stream, but we compensate by not         advancing the stream when gettok points to a token from earlier in the         stream than the one currently pointed to by peektok).
Added _very_ rudimentary support for recognizing function definitions (not
        declarations/prototypes). The parser can detect function definitions
that operate on non-pointer integral types (also must be non-void). Although
we technically do parse an arbitrary number of parameters in the parameter
list, when we install the function in the symbol table, only the first
parameter is shown for the parameter types. This is likely the next thing I
will fix. Also, since there is no support for statement parsing or local
variable declaration recognition yet, the parser just skips everything in
between the { }. Finally, the only result of parsing a function definition is
that the function will be installed in the symbol table. No parse trees are
constructed just yet.

Some clean up was done to move functions that operate purely on tokens, such as
reserved() and operator(), were moved into the same file as the other token
operations in order to clean up the parser.c file.