Several things happened during this last changeset:
- Added basic functions for creating a parse tree while we are parsing.
- Added a header for pprint.c so that we can call ppexpr which "pretty-prints"
the parse trees (for debugging purposes).
- Added code that constructs the proper portions of the parse trees in each of
the various expression productions.
- Modified the lexer to recognize "*" as a "STAR" token, which then gets
resolved to either a "DEREFERENCE" or a "MULTIPLICATION" at parse-time.
- Modified the lexer to recognize "&" as an "AMPERSAND" token, which then gets
resolved to either a "REFERENCE" or a "BITWISE_AND" token at parse-time.
- Fixed the <assignment-expression> so that it can now do both
<conditional-expressions> as well as actuall assignments.
- Added helper boolean functions to the token API to test if a token is a
certain type of token (for added code clarity in the parser).

NOTE: I had to ditch the purely recursive approach to parsing
<additive-expressions> and <multiplicative-expressions> because doing so would
cause all of the operators to become right-associative, which is ok for '+'
and '*' since they are associative operations, but '-' and '/' are
non-associative, so switching their associativity will yield an incorrect
result. Instead, we now loop inside of those functions until there are no more
operands to parse.